Adding upload logo on profil company

+++ app/Models/Profil.php
@@ -14,6 +14,7 @@ class Profil extends Model
         'name_company',
         'address',
         'phone',
+        'logo',
         'photo',
         'description'
     ];
@@ -45,16 +46,21 @@ class Profil extends Model
         // Event Lisener
         static::saving(function ($profil) {
             self::resizePhotoIfNeeded($profil);
+            self::resizeLogoIfNeeded($profil);
         });
 
         static::updating(function ($profil) {
             if ($profil->isDirty('photo')) {
                 self::deletePhotoFile($profil->getOriginal('photo'));
             }
+            if ($profil->isDirty('logo')) {
+                self::deleteLogoFile($profil->getOriginal('logo'));
+            }
         });
 
         static::deleting(function ($profil) {
             self::deletePhotoFile($profil->photo);
+            self::deleteLogoFile($profil->logo);
         });
     }
 
@@ -94,6 +100,42 @@ class Profil extends Model
         }
     }
 
+    // Resize ukuran logo
+    protected static function resizeLogoIfNeeded($profil)
+    {
+        //Cek di field input apakah ada logo
+        if (!$profil->logo) {
+            return;
+        }
+
+        //Lokasi logo
+        $fileLocation = storage_path('app/public/' . $profil->logo);
+
+        //Cek apakah ada logo di lokasi penyimpanan
+        if (!file_exists($fileLocation)) {
+            return;
+        }
+
+        $maxFileSize = 1024 * 1024; //1Mb
+
+        //Ukuran file dibawah 1Mb tidak dilakukan proses resize
+        if (filesize($fileLocation) <= $maxFileSize) {
+            return;
+        }
+
+        //Proses resize
+        $manager = new ImageManager(new Driver());
+        $logo = $manager->read($fileLocation);
+        $logo->scale(width: 800);
+
+        $quality = 80;
+        while (filesize($fileLocation) > $maxFileSize && $quality >= 30) {
+            $logo->save($fileLocation, quality: $quality);
+            clearstatcache(true, $fileLocation);
+            $quality -= 5;
+        }
+    }
+
     //Hapus foto dari storage
     protected static function deletePhotoFile($photo)
     {
@@ -103,6 +145,20 @@ class Profil extends Model
         }
 
         $fileLocation = storage_path('app/public/' . $photo);
+        if (file_exists($fileLocation)) {
+            unlink($fileLocation);
+        }
+    }
+
+    //Hapus logo dari storage
+    protected static function deleteLogoFile($logo)
+    {
+        //Cek apakah ada logo
+        if (!$logo) {
+            return;
+        }
+
+        $fileLocation = storage_path('app/public/' . $logo);
         if (file_exists($fileLocation)) {
             unlink($fileLocation);
         }
